+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Extract op and description while converting opentelemetry spans to sentry spans. ([#3287](https://github.com/getsentry/relay/pull/3287))
 - Drop `event_id` and `remote_addr` from all outcomes. ([#3319](https://github.com/getsentry/relay/pull/3319))
 - Support for AI token metrics ([#3250](https://github.com/getsentry/relay/pull/3250))
+- Accept integers in `event.user.username`. ([#3328](https://github.com/getsentry/relay/pull/3328))
 
 **Internal**: